+++ smartcontract/programs/doublezero-serviceability/src/processors/user/delete.rs
@@ -124,6 +124,10 @@ pub fn process_delete_user(
         try_acc_write(&accesspass, accesspass_account, payer_account, accounts)?;
     }
 
+    if user.status != UserStatus::Activated && user.status != UserStatus::Suspended {
+        return Err(DoubleZeroError::InvalidStatus.into());
+    }
+
     user.status = UserStatus::Deleting;
 
     try_acc_write(&user, user_account, payer_account, accounts)?;